LayoutTests/http/tests/security/contentSecurityPolicy/directive-parsing-01-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'allow'.

CONSOLE MESSAGE: Refused to load script from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because of Content-Security-Policy.

This script should not execute even though there are parse errors in the policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
